fix: popView 수정 완료
